refactor: extract implementation
